evaluator refactor
