+++ requirements.txt
@@ -2,8 +2,14 @@
 torch==1.10.0
 torchvision==0.11.1
 pytorch-lightning==1.5.1
+torchmetrics==0.6.0
 hydra-core==1.1.1
+omegaconf==2.1.1
 
+# Model hub
+requests==2.26.0
 
 # Tests
 pytest==6.2.5
+
+setuptools==57.0.0
