--- src/ulib/configs/callbacks/default.yaml
@@ -1,11 +1,0 @@
-early_stopping:
-  _target_: pytorch_lightning.callbacks.early_stopping.EarlyStopping
-  patience: 25
-  monitor: val_loss
-  mode: min
-  min_delta: 0.0001
-  verbose: True
-
-pbar:
-  _target_: pytorch_lightning.callbacks.TQDMProgressBar
-  refresh_rate: 20
