--- src/ulib/configs/datamodule/stl10_module.yaml
@@ -1,5 +1,0 @@
-_target_: ulib.data.stl10.STL10DataModule
-
-data_dir: ${data_dir}
-batch_size: 32
-num_workers: 4
