--- src/ulib/configs/trainer/trainer.yaml
@@ -1,4 +1,0 @@
-_target_: pytorch_lightning.Trainer
-
-max_epochs: 300
-gpus: 1
