+++ src/unsupervised_pretraining/configs/callbacks/evaluator_callbacks.yaml
@@ -1,0 +1,11 @@
+early_stopping:
+  _target_: pytorch_lightning.callbacks.early_stopping.EarlyStopping
+  patience: 25
+  monitor: val_loss
+  mode: min
+  min_delta: 0.0001
+  verbose: True
+
+pbar:
+  _target_: pytorch_lightning.callbacks.TQDMProgressBar
+  refresh_rate: 20
