+++ src/unsupervised_pretraining/configs/trainer/trainer.yaml
@@ -1,0 +1,4 @@
+_target_: pytorch_lightning.Trainer
+
+max_epochs: 300
+gpus: 1
